feat vendorのテストが動くように修正
