+++ backend/app/Http/Middleware/RedirectIfAuthenticated.php
@@ -2,12 +2,10 @@
 
 namespace App\Http\Middleware;
 
-use App\Models\User_ip;
 use App\Providers\RouteServiceProvider;
 use Closure;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use GeoIP;
 
 class RedirectIfAuthenticated
 {
@@ -15,23 +13,19 @@ class RedirectIfAuthenticated
      * Handle an incoming request.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \Closure  $next
+     * @param  \Closure(\Illuminate\Http\Request): (\Illuminate\Http\Response|\Illuminate\Http\RedirectResponse)  $next
      * @param  string|null  ...$guards
-     * @return mixed
+     * @return \Illuminate\Http\Response|\Illuminate\Http\RedirectResponse
      */
     public function handle(Request $request, Closure $next, ...$guards)
     {
-
-
         $guards = empty($guards) ? [null] : $guards;
-        foreach ($guards as $guard) {
 
+        foreach ($guards as $guard) {
             if (Auth::guard($guard)->check()) {
                 return redirect(RouteServiceProvider::HOME);
             }
-
         }
-
 
         return $next($request);
     }
