+++ backend/tests/Feature/Auth/AuthenticationTest.php
@@ -24,7 +24,7 @@ class AuthenticationTest extends TestCase
 
         $response = $this->post('/login', [
             'email' => $user->email,
-            'password' => 'password',
+            'password' => $user->email,
         ]);
 
         $this->assertAuthenticated();
