+++ backend/tests/Feature/Auth/PasswordConfirmationTest.php
@@ -14,8 +14,8 @@ class PasswordConfirmationTest extends TestCase
     public function test_confirm_password_screen_can_be_rendered()
     {
         $user = Features::hasTeamFeatures()
-                        ? User::factory()->withPersonalTeam()->create()
-                        : User::factory()->create();
+            ? User::factory()->withPersonalTeam()->create()
+            : User::factory()->create();
 
         $response = $this->actingAs($user)->get('/user/confirm-password');
 
@@ -27,7 +27,7 @@ class PasswordConfirmationTest extends TestCase
         $user = User::factory()->create();
 
         $response = $this->actingAs($user)->post('/user/confirm-password', [
-            'password' => 'password',
+            'password' => $user->email,
         ]);
 
         $response->assertRedirect();
