+++ backend/tests/Feature/Auth/ProfileInformationTest.php
@@ -12,8 +12,15 @@ class ProfileInformationTest extends TestCase
 {
     use RefreshDatabase;
 
+    /**
+     * jetstreamデフォルトのlivewire使ってないのでこのテストは絶対に失敗する
+     *
+     * @return void
+     */
     public function test_current_profile_information_is_available()
     {
+        return $this->markTestSkipped('Livewireでは実装していないため省略');
+
         $this->actingAs($user = User::factory()->create());
 
         $component = Livewire::test(UpdateProfileInformationForm::class);
@@ -24,11 +31,13 @@ class ProfileInformationTest extends TestCase
 
     public function test_profile_information_can_be_updated()
     {
+        return $this->markTestSkipped('Livewireでは実装していないため省略');
+
         $this->actingAs($user = User::factory()->create());
 
         Livewire::test(UpdateProfileInformationForm::class)
-                ->set('state', ['name' => 'Test Name', 'email' => 'test@example.com'])
-                ->call('updateProfileInformation');
+            ->set('state', ['name' => 'Test Name', 'email' => 'test@example.com'])
+            ->call('updateProfileInformation');
 
         $this->assertEquals('Test Name', $user->fresh()->name);
         $this->assertEquals('test@example.com', $user->fresh()->email);
